Add files via upload
Hey there, this an Excel project of 2011 Census Report. So, here I make two dashboards which allow you to gather information about the report you collect information of education, literacy rate, worker type, cast, gender, age group along with that I make a slicer on the basis of which allow you to easily access of reports on district level. And in another dashboard, I charted an Indian map which provide information of population of state with the help of using heatmap (if color of state is darker it means that their population is higher and vice-versa) and I make some cards which tell total population, total male, total female, literacy rate and draw buttons also to easily access both of the dashboards.